Share a portion of LLC with the Vulcano model (#9)
* cleanup: remove unnecessary components from vul_model board

Now that we have a working model and know exactly what we need, it's easier
to perform this cleanup pass and remove unnecessary stuff.

* move ZMQ-specific code to utils

We want to use ZMQ to send both register reads/writes to the model as well
as perform standard memory accesses. Move it to a shared place so that
both the board implementation and the CSR memory map can use it.

* add ZMQ support for reading and writing memory

* add new device to implement Vulcano memory

This device forwards memory access happening within the DDR
memory region to the Vulcano model, so that we only have one
source of truth for DDR. This is required for config space to
work, as the HW uses memory provided by FW.

* add to the Vulcano board the new memory device

Because of low performance when running from the model's memory,
only share the portion where we put the memory shared with the
HW.

* exit the simulation on read/write errors from the model