simbridge: New option to pass type 0 config to model/FW for upstream bridge.
Vulcano needs to handle type 0 config for the upstream bridge either
partially (PCIe mode) or completely (UALink mode) in firmware.  This
change allows this behaviour to be requested by setting the
SIMBRIDGE_VULCANO_TYPE0 environment variable.

(The model at the other end needs to be recent enough to provide the
support for this as well, of course.)

Behaviour is unchanged by default so this does not affect use for previous
ASICs.