Add a "model accelerator" device to the Vulcano model board (#11)
* add support for mem operations bigger than VUL_ZMQ_BUF_SIZE size

* support CSR writes batching

* Add vul_zmq APIs to reset memory, step doorbells and get the max transmit size

* fix csr wr

* Introduce the model accelerator device

It is a device for which we'll write a Zephyr device.
This device is capable of combining what would normally
be multiple operations in a single message to optimise
modeling performance.

* Add modacc device to the Vulcano model board

* Bump SRAM size to appease QEMU when loading bigger FW images

Although we're loading FW in LLC, the SRAM is mapped
to the "main" memory created by command line options
by QEMU. QEMU uses this memory's size to determine
whether FW can be loaded or not due to size constraints.
To load images bigger than 512 KiB without hacking
too much into QEMU internals, we bump the SRAM area
to be as big as the LLC itself. We have address space
to spare in that region so this is not an issue.
The main benefit in doing this, rather than just removing
this region and use straight away the LLC, is that
this way we can still partition the 32 MiB of the LLC
with the nicram region without having the LLC region
mapped to plain RAM bigger than what it should be.
This is desirable to catch OOB accesses made by FW.