VSW-286: HACK: force simbridge to handle extended config space
For reasons not completely understood, something is going wrong
in QEMU's object model and the correct realise function setting
up the PCI_BUS_EXTENDED_CONFIG_SPACE for a downstream port
isn't being called. Because of this, QEMU autonomously replies
with all 1s when host software tries to access extended
capabilities of an endpoint implemented in the model, without
ever trying to invoke the simbridge code.

This workaround is dirty, but unblocks us as it forces QEMU
to forward us extended config space access.