VSW-306: Implement sockdma device
Create nonblocking socket to pass data between firmware running on qemu
and host software. On 'read_status' register read, poll the socket for
rx data and write it to a memory mapped region accessible by firmware.
Sends read data populated in this region by the firmare and send it
through the socket.